first time setup fix

+++ setup.sh
@@ -12,13 +12,28 @@ sudo chown -R piuser:piuser $VIDEOS_DIR
 echo "Installing ffmpeg and python libraries..."
 sudo apt install -y ffmpeg python3-opencv python3-numpy python3-dotenv
 
-echo "Stopping/disabling old services..." #todo remove this eventually #todo deal with this on the first run?
-sudo systemctl disable capture.service
-sudo systemctl stop capture.service
-sudo systemctl disable motion_postprocess.service
-sudo systemctl stop motion_postprocess.service
-sudo systemctl disable clip_uploader.service
-sudo systemctl stop clip_uploader.service
+echo "Stopping/disabling old services..."
+# Function to check if a service exists
+service_exists() {
+    systemctl list-unit-files "$1" &>/dev/null
+    return $?
+}
+
+# Stop and disable services if they exist
+if service_exists "capture.service"; then
+    sudo systemctl stop capture.service
+    sudo systemctl disable capture.service
+fi
+
+if service_exists "motion_postprocess.service"; then
+    sudo systemctl stop motion_postprocess.service
+    sudo systemctl disable motion_postprocess.service
+fi
+
+if service_exists "clip_uploader.service"; then
+    sudo systemctl stop clip_uploader.service
+    sudo systemctl disable clip_uploader.service
+fi
 
 echo "Installing systemd services..."
 sudo cp /home/piuser/rpi-surveillence/capture.service /etc/systemd/system/capture.service
